pyproject.toml
[project]
name = "fa2025-cs511-mp2-hnsw"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.12.0",
    "h5py>=3.15.1",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
